+++ CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.16.0] - 2018-06-03
+
+### Added
+
+- Optionally reference the todo.txt filename from which the next actions were read. Closes #38.
 
 ### Changed
 
-- Optionally reference the todo.txt filename from which the next actions were read. Closes #38.
 - Reorganized the help information.
 
 ## [0.15.0] - 2018-06-02